#22: Parse energy points in the future
The original code was breaking one of my rules which is that you should only increase indentation (e.g. using an if) only when something goes wrong.

The happy path here is that we're going to be parsing EnergyPoints, so instead of using an if, `filter` the results first.

Also note how we're using `let` instead of `var` now. That adds immutability to the array which is always a good thing.