#22: Map results with rounded data.
Now that we have `EnergyPoint`s that are interesting to us, map them one by one by transforming their data to something the app will consume (i.e. rounded numbers).

The documentation is pretty clear that we should not be using `.sign` to determine a floating point's negativity. My code uses more lines but it is more readable and tests negativity numerically.