#22: Complete migration
Finish the `.map` by returning the new EnergyPricePoint with the rounded marketprice.

Also, now that we have a filtered array, min and max are found in a single line each. Note that we are indeed iterating the array more times, but you should _always_ make your code readable first and optimize by measuring. If these changes are substantial, we will find out when we run Instruments during QA or code review (if your reviewer is super thorough).